Moves output and reduces scope
        Output object grabbed options to determine limiting, but I
decided to leave that in the bigger block for now and hope it finds a
better home. Also, move the Output object to its own file.
